Changed the onlyif on rmdir

+++ manifests/make.pp
@@ -107,6 +107,7 @@ define drush::make ($makefile,
 
   exec { "drush-make-rmdir-${makefile}-${build_path}":
     command  => "rmdir ${build_path} &> /dev/null || true",
+    onlyif => "test -d ${build_path}"
   }
 
   # Since drush make won't overwrite an existing directory, we remove an empty directory if it's there already (say for instance puppet/apache already provisioned it)
